Update per latest 2 comments
